Moved contexts to their own files.  ALSO I THINK THIS IS WORKING NOW :D
