+++ client/src/components/home/shared/AskQuestion.tsx
@@ -8,8 +8,9 @@ import CooldownViolationOverlay from './CooldownViolationOverlay';
 import BaseCard from '../../common/cards/BaseCard';
 
 import HomeService from '../../../services/HomeService';
-import SettingsService from '../../../services/SettingsService';
-import {QueueDataContext, StudentDataContext, UserDataContext} from '../../../App';
+import {UserDataContext} from '../../../contexts/UserDataContext';
+import {QueueDataContext} from '../../../contexts/QueueDataContext';
+import {StudentDataContext} from '../../../contexts/StudentDataContext';
 
 function createData(assignment_id, name) {
   return {assignment_id, name};
@@ -20,7 +21,6 @@ const date = new Date();
 export default function AskQuestion() {
   const {userData} = useContext(UserDataContext);
   const {queueData} = useContext(QueueDataContext);
-
   // not changing name or andrewID to use global because this component can also be used by TAs to manually add questions
   const [name, setName] = useState('');
   const [andrewID, setAndrewID] = useState('');
