+++ client/src/components/home/shared/CooldownViolationOverlay.tsx
@@ -4,14 +4,13 @@ import {
 } from '@mui/material';
 
 import HomeService from '../../../services/HomeService';
-import {QueueDataContext} from '../../../App';
+import {QueueDataContext} from '../../../contexts/QueueDataContext';
 
 export default function CooldownViolationOverlay(props) {
   const {open, setOpen, andrewID, question, location, topic, timePassed} = props;
   const theme = useTheme();
 
   const {queueData} = useContext(QueueDataContext);
-
   function callAddQuestionAPIOverrideCooldown() {
     HomeService.addQuestion(
         JSON.stringify({
