+++ client/src/components/home/student/StudentMain.tsx
@@ -10,7 +10,9 @@ import AskQuestion from '../shared/AskQuestion';
 import HomeService from '../../../services/HomeService';
 import {StudentStatusValues} from '../../../services/StudentStatus';
 import {socketSubscribeTo, socketUnsubscribeFrom} from '../../../services/SocketsService';
-import {QueueDataContext, StudentDataContext, UserDataContext} from '../../../App';
+import {UserDataContext} from '../../../contexts/UserDataContext';
+import {QueueDataContext} from '../../../contexts/QueueDataContext';
+import {StudentDataContext} from '../../../contexts/StudentDataContext';
 
 function StudentMain() {
   // const [questionValue, setQuestionValue] = useState('');
