+++ client/src/components/settings/VideoChatSettings.tsx
@@ -7,7 +7,7 @@ import {
 import BaseCard from '../common/cards/BaseCard';
 
 import SettingsService from '../../services/SettingsService';
-import {UserDataContext} from '../../App';
+import {UserDataContext} from '../../contexts/UserDataContext';
 
 export default function VideoChatSettings(props) {
   const {userData} = useContext(UserDataContext);
